Refactor QuestionService to use DTOs for data transfer

+++ backend/src/main/java/com/quiz/backend/service/QuestionServiceImpl.java
@@ -1,17 +1,20 @@
 package com.quiz.backend.service;
 
+import com.quiz.backend.dto.QuestionDTO;
 import com.quiz.backend.entity.Choice;
 import com.quiz.backend.entity.Question;
 import com.quiz.backend.entity.Quiz;
+import com.quiz.backend.exception.QuestionNotFoundException;
 import com.quiz.backend.exception.QuizNotFoundException;
 import com.quiz.backend.repository.ChoiceRepository;
 import com.quiz.backend.repository.QuestionRepository;
 import com.quiz.backend.repository.QuizRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class QuestionServiceImpl implements QuestionService {
@@ -19,28 +22,38 @@ public class QuestionServiceImpl implements QuestionService {
     private final QuestionRepository questionRepository;
     private final QuizRepository quizRepository;
     private final ChoiceRepository choiceRepository;
+    private final ModelMapper modelMapper;
 
-    public QuestionServiceImpl(QuestionRepository questionRepository, QuizRepository quizRepository, ChoiceRepository choiceRepository) {
+    public QuestionServiceImpl(QuestionRepository questionRepository, QuizRepository quizRepository, ChoiceRepository choiceRepository, ModelMapper modelMapper) {
         this.questionRepository = questionRepository;
         this.quizRepository = quizRepository;
         this.choiceRepository = choiceRepository;
+        this.modelMapper = modelMapper;
     }
 
     @Override
-    public List<Question> getAllQuestion(Long quizId, Long userId) {
-        return questionRepository.findByQuiz_IdAndQuiz_User_Id(quizId, userId);
+    public List<QuestionDTO> getAllQuestion(Long quizId, Long userId) {
+        List<Question> questions=  questionRepository.findByQuiz_IdAndQuiz_User_Id(quizId, userId);
+
+        return questions.stream()
+                .map(question -> modelMapper.map(question, QuestionDTO.class))
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Question saveQuestion(Long quizId, Long userId, Question newQuestion) {
+    public Question saveQuestion(Long quizId, Long userId, QuestionDTO newQuestion) {
 
         Quiz quiz = quizRepository.findByIdAndUser_Id(quizId, userId).orElseThrow(() -> new QuizNotFoundException("Quiz Not Found"));
-        newQuestion.setQuiz(quiz);
-        return questionRepository.save(newQuestion);
+        Question question = modelMapper.map(newQuestion, Question.class);
+        question.setCreatedAt(LocalDateTime.now());
+        question.setUpdatedAt(LocalDateTime.now());
+        question.setVersion(1);
+        question.setQuiz(quiz);
+        return questionRepository.save(question);
     }
 
     @Override
-    public Question updateQuestion(Long userId, Long quizId, Long questionId, Question question) {
+    public Question updateQuestion(Long userId, Long quizId, Long questionId, QuestionDTO question) {
         Question existingQuestion = questionRepository.findByQuiz_IdAndQuiz_User_IdAndId(quizId, userId, questionId);
         existingQuestion.setContent(question.getContent());
         existingQuestion.setUpdatedAt(LocalDateTime.now());
@@ -51,12 +64,15 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     public void deleteQuestion(Long userId, Long quizId, Long questionId) {
         Question question = questionRepository.findByQuiz_IdAndQuiz_User_IdAndId(quizId, userId, questionId);
+        if (question == null) {
+            throw new QuestionNotFoundException("Question not found");
+        }
 
-        //delete all choices associated with the question
-        List<Choice> choice = choiceRepository.findByQuestion_Id(questionId);
-        choiceRepository.deleteByQuestion_Id(questionId);
+        // Delete all choices associated with the question
+        List<Choice> choices = choiceRepository.findByQuestion_Id(questionId);
+        choiceRepository.deleteAll(choices);
 
-        //delete the question
+        // Delete the question
         questionRepository.delete(question);
     }
 }
